add `PtrToBytes`
Convinience function for custom widgets. This will convert the literal pointer
(NOT the data it points to) to a byte slice (to be used with GetID()). This is
basically equivalent to passing a pointer pointer to mu_get_id.

This function will allocate a new byte slice on the heap. It produces different
results based on the endianess of the machine, but that should be fine for hashing purposes.